modify all int to int64 (#366)

+++ pkg/deviceshifu/deviceshifubase/deviceshifubase_test.go
@@ -125,7 +125,7 @@ func TestStartTelemetryCollection(t *testing.T) {
 								DeviceShifuTelemetryPushToServer:      unitest.ToPointer(false),
 								DeviceShifuTelemetryCollectionService: unitest.ToPointer("test_endpoint-1"),
 							},
-							InitialDelayMs: unitest.ToPointer(1),
+							InitialDelayMs: unitest.ToPointer(int64(1)),
 						},
 					},
 				},
