+++ pkg/deviceshifu/deviceshifusocket/deviceshifusocketconfig.go
@@ -3,13 +3,13 @@ package deviceshifusocket
 // RequestBody Socket Request Body
 type RequestBody struct {
 	Command string `json:"command"`
-	Timeout int    `json:"timeout"`
+	Timeout int64  `json:"timeout"`
 }
 
 // ReturnBody Socket Reply Body
 type ReturnBody struct {
 	Message string `json:"message"`
-	Status  int    `json:"status"`
+	Status  int64  `json:"status"`
 }
 
 const (
